Sharing Test Context between Cucumber Step Definitions
